Relocated ChainComparator The most complex algorithms of ChainComparator commented
