+++ fwoptimizer/classes/fdd.py
@@ -1296,370 +1296,4 @@ class FDD:
             if option1.isDisjoint(option2):  # Check if they have any common elements
                 return True
         return False
-
-
-class ChainComparator:
-
-    class PseudoRule:
-
-        def __init__(self) -> None:
-
-            self._id = None
-            self._decision = None
-            self._predicates = {}
-
-        def __repr__(self) -> str:
-            """
-            Rule __repr__
-            
-            Returns:
-                str: PseudoRule String representation
-            """
-            return f"PseudoRule {self._id}: {self._predicates} -> {self._decision}"
-
-        def fillFromRule(self, rule: Rule, fieldList: FieldList):
-
-            self._id = rule.getId()
-            self._decision = rule.getDecision()
-            self._predicates = {}
-
-            for field in fieldList.getFields():
-                element_list = rule.getOption(field.getName(), None)
-                self._predicates[field.getName()] = ElementSet.createElementSet(field.getType(), element_list if element_list else [])
-        
-        def setId(self, id):
-            self._id = id
-
-        def getId(self):
-            return self._id
-
-        def setDecision(self, decision):
-            self._decision = decision
-
-        def getDecision(self):
-            return self._decision
-        
-        def setPredicate(self, key, value):
-            self._predicates[key] = value
-
-        def getPredicates(self):
-            return self._predicates
-        
-        def replicate(self):
-            rep = ChainComparator.PseudoRule()
-            rep.setId(self.getId())
-            rep.setDecision(self.getDecision())
-            for key in self.getPredicates():
-                rep.setPredicate(key, self.getPredicates()[key].replicate())
-            return rep
-        
-        def sameFieldValues(self, other: "ChainComparator.PseudoRule"):
-            """summary
-            """
-            for key in self.getPredicates():
-                if self.getPredicates()[key] != other.getPredicates()[key]:
-                    return False
-            return True
-
-        def isNull(self):
-
-            for key in self.getPredicates():
-                if self.getPredicates()[key].isEmpty():
-                    return True
-            return False
-
-        def intersection(self, other: "ChainComparator.PseudoRule", fieldList: FieldList):
-
-            result = ChainComparator.PseudoRule()
-
-            for field in fieldList.getFields():
-                
-                option1 = self.getPredicates().get(field.getName())
-                option2 = other.getPredicates().get(field.getName())
-
-                intersection = option1.intersectionSet(option2)
-
-                result.setPredicate(field.getName(), intersection)
-            
-            if result.isNull():
-                return None
-            return result
-        
-        def difference(self, other: "ChainComparator.PseudoRule", fieldList: FieldList) -> List[Rule]:
-                
-            diff = []
-
-            for field in fieldList.getFields():
-
-                fDiff = self.getPredicates()[field.getName()].differenceSet(other.getPredicates()[field.getName()])
-                if not fDiff.isEmpty():
-                    rule = self.replicate()
-                    rule.setPredicate(field.getName(), fDiff)
-                    diff.append(rule)
-
-            i = 0
-            while i < len(diff)-1:
-
-                intersection = diff[i].intersection(diff[i+1], fieldList)
-
-                if intersection != None:
-
-                    if intersection.sameFieldValues(diff[i]):
-                        diff.pop(i)
-                    elif intersection.sameFieldValues(diff[i+1]):
-                        diff.pop(i+1)
-                    else:
-                        i += 1
-
-                else:
-                    i += 1
-
-            return diff
-    
-    class PseudoChain:
-
-        def __init__(self) -> None:
-            self._name = None
-            self._defaultdecision = None
-            self._rules = []
-
-        def __repr__(self) -> str:
-            """
-            Chain __repr__
-
-            Returns:
-                str: Chain String representation
-            """
-            rules_str = "\n".join([str(rule) for rule in self._rules])
-            return f"{self._name}:\n{rules_str if rules_str else ''}"
-        
-        def fillFromChain(self, chain: Chain, fieldList: FieldList):
-            self._name = chain.getName()
-            self._defaultdecision = chain.getDefaultDecision()
-            self._rules = []
-
-            for rule in chain.getRules():
-                aux = ChainComparator.PseudoRule()
-                aux.fillFromRule(rule, fieldList)
-                self._rules.append(aux)
-
-        def addRule(self, rule):
-            self._rules.append(rule)
-
-        def getRules(self):
-            return self._rules
-
-        def setName(self, name):
-            self._name = name
-        
-        def getName(self):
-            return self._name
-        
-        def setDefaultDecision(self, defaultDecision):
-            self._defaultdecision = defaultDecision
-        
-        def getDefaultDecision(self):
-            return self._defaultdecision
-        
-        def replicate(self):
-
-            rep = ChainComparator.PseudoChain()
-            rep.setName(self.getName())
-            rep.setDefaultDecision(self.getDefaultDecision())
-            for rule in self.getRules():
-                rep.addRule(rule.replicate())
-            return rep
-        
-        def getEffectiveChain(self, fieldList: FieldList) -> "ChainComparator.PseudoChain":
-
-            if self.getDefaultDecision() != "DROP":
-                raise ValueError(f"La cadena tiene Default decision {self.getDefaultDecision()} y por el momento solo se pueden comparar cadenas con default decision DROP")
-
-            newChain = ChainComparator.PseudoChain()
-            newChain.setDefaultDecision(self.getDefaultDecision())
-            newChain.setName(f"effective-{self.getName()}")
-
-            rules = self.getRules()
-
-            for i in range(len(rules)):
-
-                print(f"\ni: {i}")
-                print(rules[i].getDecision())
-
-                if rules[i].getDecision() == "ACCEPT":
-
-                    resRules = [rules[i]]
-
-                    for j in range(i):
-
-                        print(f"\nj: {j}")
-                        print(f"resRules: {resRules}")
-
-                        resRules2 = []
-
-                        for k in range(len(resRules)):
-
-                            print(f"\nk: {k}")
-
-                            diff = resRules[k].difference(rules[j], fieldList)
-
-                            print(f"restando A-B\nA: {resRules[k]}\nB: {rules[j]}")
-                            print(f"diff: {diff}")
-
-                            if diff != None:
-
-                                resRules2 += diff
-
-                        i = 0
-                        while i < len(resRules2)-1:
-
-                            intersection = resRules2[i].intersection(resRules2[i+1], fieldList)
-
-                            if intersection != None:
-
-                                if intersection.sameFieldValues(resRules2[i]):
-                                    resRules2.pop(i)
-                                elif intersection.sameFieldValues(resRules2[i+1]):
-                                    resRules2.pop(i+1)
-                                else:
-                                    i += 1
-
-                            else:
-                                i += 1
-
-                        resRules = resRules2
-
-                    for rule in resRules:
-
-                        print(f"Agregando a newChain: {rule}")
-                        newChain.addRule(rule)
-                
-            return newChain
-
-
-    def __init__(self, fieldList: FieldList) -> None:
-        
-        self._fieldList = fieldList
-        self._chain1 = None
-        self._chain2 = None
-        self._effectiveChain1 = None
-        self._effectiveChain2 = None
-    
-    def __repr__(self) -> str:
-        
-        return f"CHAIN COMPARATOR\n\nORIGINALS\n\n{self._chain1}\n\n{self._chain2}\
-                \n\nCONVERTED\n\n{self._effectiveChain1}\n\n{self._effectiveChain2}"
-    
-    def setFieldList(self, fieldList: FieldList):
-        self._fieldList = fieldList
-    
-    def setChain1FromChain(self, chain1: Chain):
-        self._chain1 = ChainComparator.PseudoChain()
-        self._chain1.fillFromChain(chain1, self._fieldList)
-        self._effectiveChain1 = self._chain1.getEffectiveChain(self._fieldList)
-
-    def setChain2FromChain(self, chain2: Chain):
-        self._chain2 = ChainComparator.PseudoChain()
-        self._chain2.fillFromChain(chain2, self._fieldList)
-        self._effectiveChain2 = self._chain2.getEffectiveChain(self._fieldList)
-
-    def areEquivalents(self):
-
-        if self._effectiveChain1 and self._effectiveChain2:
-
-            #Comparacion 1 contra 2
-
-            for rule1 in self._effectiveChain1.getRules():
-
-                residual = [rule1]
-
-                for rule2 in self._effectiveChain2.getRules():
-
-                    accum = []
-
-                    for rule in residual:
-
-                        diff = rule.difference(rule2, self._fieldList)
-
-                        if diff != None:
-
-                            accum += diff
-
-                    i = 0
-                    while i < len(accum)-1:
-
-                        intersection = accum[i].intersection(accum[i+1], self._fieldList)
-
-                        if intersection != None:
-
-                            if intersection.sameFieldValues(accum[i]):
-                                accum.pop(i)
-                            elif intersection.sameFieldValues(accum[i+1]):
-                                accum.pop(i+1)
-                            else:
-                                i += 1
-
-                        else:
-                            i += 1
-
-                    residual = accum
-
-                for result in residual:
-
-                    if not result.isNull():
-                        return False
-
-            #Comparacion 2 contra 1
-
-            for rule1 in self._effectiveChain2.getRules():
-
-                residual = [rule1]
-
-                for rule2 in self._effectiveChain1.getRules():
-
-                    accum = []
-
-                    for rule in residual:
-
-                        diff = rule.difference(rule2, self._fieldList)
-
-                        if diff != None:
-
-                            accum += diff
-
-                    i = 0
-                    while i < len(accum)-1:
-
-                        intersection = accum[i].intersection(accum[i+1], self._fieldList)
-
-                        if intersection != None:
-
-                            if intersection.sameFieldValues(accum[i]):
-                                accum.pop(i)
-                            elif intersection.sameFieldValues(accum[i+1]):
-                                accum.pop(i+1)
-                            else:
-                                i += 1
-
-                        else:
-                            i += 1
-
-                    residual = accum
-
-                for result in residual:
-
-                    if not result.isNull():
-                        return False
-                    
-            # Si hasta aqui no retornamos un falso, entonces es True
-
-            return True
-        
-        else:
-
-            raise ValueError("Debe cargar primero dos Chain en ChainComparator")
-
-            
-                    
-
