+++ fwoptimizer/utils/__init__.py
@@ -4,3 +4,5 @@ utils Package
 
 # AliasDict module
 from .aliasDict import AliasDefaultDict
+
+from .chainComparator import ChainComparator
